==> docs/03_full-df.qmd <==
---
title: '03 Final Dataframe'
author:
  - name: J Andrés Gannon
    email: "juan.gannon@vanderbilt.edu"
    url: jandresgannon.com
    affiliation: Vanderbilt University
format:
  pdf:
    fig-format: png
    thanks: test text here
editor: visual
execute:
  echo: false
  warning: false
  cache: true
date: "`r format(Sys.Date(), '%B %d, %Y')`"
header-includes:
- \usepackage{tikz}
- \usepackage{pgfplots}
- \pgfplotsset{compat=newest}
- \usetikzlibrary{plotmarks}
- \usetikzlibrary{arrows.meta}
- \usepgfplotslibrary{patchplots}
- \usepackage{grffile}
- \usepackage{caption}
- \usepackage[utf8]{inputenc}
- \usepackage[doublespacing]{setspace}
- \AtBeginEnvironment{tabular}{\singlespacing}
- \usepackage{float}
- \usepackage{multirow}
- \usepackage{tablefootnote}
- \usepackage{pifont}
- \usepackage{newunicodechar}
- \usepackage{booktabs}
- \usepackage{tabularx}
- \newunicodechar{✓}{\ding{51}}
bibliography: ForceStructure.bib
---

This document creates the full dataframe used for all models in the paper

# Prep

Pipe operators have trouble loading in-line so we load those first.

```{r}
library(magrittr)
library(ggplot2)
```

# Load and clean data

## Load

We start by loading the cleaned IISS data. This dataframe contains information about the count of military units at the country-year level.

```{r}
# Load data, dropping IISS countryname and rename ccode var
df <- readRDS(file = paste0(here::here(), '/data/01_DV-specialization.rds')) %>%
  dplyr::select(-country) %>%
  dplyr::rename(country_code_cow = ccode)
```

## Merge with covariates

We have already created a dataframe that has the left-hand side covariates. We merge that into this dataset and then subset to the non-political variables

```{r}
# Load RHS
rhs <- readRDS(file = paste0(here::here(), '/data/02_EV.rds'))

# Merge all
df_full <- dplyr::left_join(rhs, df)

# Prep
df_full <- df_full %>%
  dplyr::mutate(dca_binary = as.factor(dca_binary),
                greatpower = as.factor(greatpower),
                democracy = as.factor(democracy),
                interstatewar_5yrlag_binary = as.factor(interstatewar_5yrlag_binary),
                country = as.factor(country),
                decade = as.factor(decade),
                year_base = (year - 1970),
                year_sq = (year - 1970)^2,
                year_cube = (year - 1970)^3) %>%
  dplyr::relocate(dplyr::starts_with("spec_"), .after = in_iiss)

# Minor cleaning
df_full <- df_full %>%
  dplyr::ungroup() %>%
  dplyr::filter(!(year == 1972 &
                    country == "Vietnam" &
                    is.na(interstatewar_5yrlag_binary))) %>%
  dplyr::distinct() %>%
  tibble::rowid_to_column(var = "id")

# Save this as the master dataframe
saveRDS(df_full, paste0(here::here(), "/data/","03_DF-full.rds"))
```
